--- .eslintrc.json
@@ -1,40 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "es2021": true
-    },
-    "extends": [
-        "airbnb-base"
-    ],
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": 12,
-        "sourceType": "module",
-        "project": "./tsconfig.json"
-    },
-    "plugins": [
-        "@typescript-eslint"
-    ],
-    "rules": {
-        "@typescript-eslint/type-annotation-spacing": ["warn", {
-            "before": false,
-            "after": true
-        }],
-        "no-redeclare":"off",
-        "prefer-destructuring":"off",
-        "no-underscore-dangle":"off",
-        "no-undef":"off",
-        "new-cap":"off",
-        "no-param-reassign":"off",
-        "class-methods-use-this":"off",
-        "no-console":"off",
-        "import/extensions": "off"
-    },
-    "settings": {
-        "import/resolver": {
-          "node": {
-            "extensions": [".js", ".ts"]
-          }
-        }
-    }
-}
