+++ .gitignore
@@ -1,7 +1,3 @@
-example/*bundle.js
-example/*bundle.js.LICENSE.txt
-example/maplibre-gl-export.*
-
 # Logs
 logs
 *.log
