+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+	"useTabs": true,
+	"singleQuote": true,
+	"trailingComma": "none",
+	"printWidth": 100,
+	"plugins": [],
+	"pluginSearchDirs": ["."]
+}
