--- lib/index.ts
@@ -1,4 +1,0 @@
-export { default as MaplibreExportControl } from './export-control';
-export {
-  Size, PageOrientation, Format, DPI,
-} from './map-generator';
