--- lib/local/de.ts
@@ -1,9 +1,0 @@
-const translation = {
-  PageSize: 'Papierformat',
-  PageOrientation: 'Papierausrichtung',
-  Format: 'Dateiformat',
-  DPI: 'Druckauflösung',
-  Generate: 'Erstellen',
-};
-
-export default translation;
