--- lib/local/en.ts
@@ -1,9 +1,0 @@
-const translation = {
-  PageSize: 'Page Size',
-  PageOrientation: 'Page Orientation',
-  Format: 'Format',
-  DPI: 'DPI',
-  Generate: 'Generate',
-};
-
-export default translation;
