--- lib/local/fr.ts
@@ -1,9 +1,0 @@
-const translation = {
-  PageSize: 'Taille de page',
-  PageOrientation: 'Orientation de la page',
-  Format: 'Format',
-  DPI: 'DPI',
-  Generate: 'Générer',
-};
-
-export default translation;
