--- lib/local/sv.ts
@@ -1,9 +1,0 @@
-const translation = {
-  PageSize: 'Sidstorlek',
-  PageOrientation: 'Sidorientering',
-  Format: 'Format',
-  DPI: 'DPI',
-  Generate: 'Generera',
-};
-
-export default translation;
