+++ package.json
@@ -1,61 +1,72 @@
 {
-  "name": "@watergis/maplibre-gl-export",
-  "version": "1.3.9",
-  "description": "This module adds control which can export PDF and images.",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "css"
-  ],
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint lib/*.ts",
-    "lint:fix": "eslint --fix lib/*.ts",
-    "build": "tsc && npm run build:cdn && npm run build:example",
-    "build:example": "webpack --mode production --config webpack.config.js",
-    "build:cdn": "webpack --mode production --config webpack.cdn.config.js && cp ./dist/cdn/maplibre-gl-export.js ./example/. && cp ./css/styles.css ./example/maplibre-gl-export.css",
-    "dev": "webpack serve --mode development --open --hot --config webpack.config.js",
-    "deploy": "gh-pages -d example -m 'deploy to gh-pages'"
-  },
-  "keywords": [
-    "mapbox",
-    "mapbox-gl-js",
-    "export",
-    "image",
-    "pdf"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/watergis/maplibre-gl-export.git"
-  },
-  "author": "Jin IGARASHI",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/watergis/maplibre-gl-export/issues"
-  },
-  "homepage": "https://github.com/watergis/maplibre-gl-export#readme",
-  "devDependencies": {
-    "@types/geojson": "^7946.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.36.2",
-    "@typescript-eslint/parser": "^5.36.2",
-    "css-loader": "^6.7.1",
-    "es6-promise": "^4.2.8",
-    "eslint": "^8.23.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "gh-pages": "^4.0.0",
-    "path": "^0.12.7",
-    "style-loader": "^3.3.1",
-    "ts-loader": "^9.3.1",
-    "typescript": "^4.8.3",
-    "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.11.0"
-  },
-  "dependencies": {
-    "js-loading-overlay": "^1.2.0",
-    "jspdf": "^2.5.1",
-    "maplibre-gl": "^2.4.0"
-  }
+	"name": "@watergis/maplibre-gl-export",
+	"version": "1.3.9",
+	"description": "This module adds control which can export PDF and images.",
+	"main": "dist/maplibre-gl-export.umd.js",
+	"types": "./dist/types/index.d.ts",
+	"module": "dist/maplibre-gl-export.es.js",
+	"exports": {
+		".": {
+			"import": "./dist/maplibre-gl-export.es.js",
+			"require": "./dist/maplibre-gl-export.umd.js"
+		}
+	},
+	"files": [
+		"dist",
+		"package.json",
+		"LICENSE"
+	],
+	"scripts": {
+		"dev": "vite",
+		"build": "npm run build:js && npm run build:css",
+		"build:js": "tsc && vite build",
+		"build:scss": "sass --no-source-map --style=compressed src/scss/maplibre-gl-export.scss:dist/maplibre-gl-export.css",
+		"build:postcss": "postcss dist/*.css -r",
+		"build:css": "npm run build:scss && npm run build:postcss",
+		"preview": "vite preview",
+		"lint": "prettier --plugin-search-dir . --check . && eslint .",
+		"format": "prettier --plugin-search-dir . --write .",
+		"deploy": "gh-pages -d dist -m 'deploy to gh-pages'"
+	},
+	"keywords": [
+		"mapbox",
+		"mapbox-gl-js",
+		"export",
+		"image",
+		"pdf"
+	],
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/watergis/maplibre-gl-export.git"
+	},
+	"author": "Jin IGARASHI",
+	"license": "MIT",
+	"bugs": {
+		"url": "https://github.com/watergis/maplibre-gl-export/issues"
+	},
+	"homepage": "https://github.com/watergis/maplibre-gl-export#readme",
+	"devDependencies": {
+		"@types/geojson": "^7946.0.10",
+		"@types/node": "^18.15.11",
+		"@typescript-eslint/eslint-plugin": "^5.36.2",
+		"@typescript-eslint/parser": "^5.36.2",
+		"autoprefixer": "^10.4.14",
+		"eslint": "^8.23.0",
+		"eslint-config-prettier": "^8.6.0",
+		"eslint-plugin-import": "^2.26.0",
+		"gh-pages": "^4.0.0",
+		"path": "^0.12.7",
+		"postcss": "^8.4.21",
+		"postcss-cli": "^10.1.0",
+		"prettier": "^2.8.4",
+		"sass": "^1.61.0",
+		"sass-loader": "^13.2.2",
+		"typescript": "^4.8.3",
+		"vite": "^4.2.1"
+	},
+	"dependencies": {
+		"js-loading-overlay": "^1.2.0",
+		"jspdf": "^2.5.1",
+		"maplibre-gl": "^2.4.0"
+	}
 }
