+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export { default as MaplibreExportControl } from './export-control';
+export { Size, PageOrientation, Format, DPI } from './map-generator';
