+++ src/lib/local/en.ts
@@ -1,0 +1,9 @@
+const translation = {
+	PageSize: 'Page Size',
+	PageOrientation: 'Page Orientation',
+	Format: 'Format',
+	DPI: 'DPI',
+	Generate: 'Generate'
+};
+
+export default translation;
