+++ src/lib/local/fi.ts
@@ -1,0 +1,9 @@
+const translation = {
+	PageSize: 'Sivukoko',
+	PageOrientation: 'Sivun suunta',
+	Format: 'Muoto',
+	DPI: 'DPI',
+	Generate: 'Generoi'
+};
+
+export default translation;
