+++ src/lib/local/fr.ts
@@ -1,0 +1,9 @@
+const translation = {
+	PageSize: 'Taille de page',
+	PageOrientation: 'Orientation de la page',
+	Format: 'Format',
+	DPI: 'DPI',
+	Generate: 'Générer'
+};
+
+export default translation;
