+++ src/lib/local/index.ts
@@ -6,13 +6,11 @@ import swedish from './sv';
 import spanish from './es';
 
 type Translation = {
-  PageSize: string;
-  PageOrientation: string;
-  Format: string;
-  DPI: string;
-  Generate: string;
-}
-
-export {
-  english, french, finnish, german, swedish, spanish, Translation,
+	PageSize: string;
+	PageOrientation: string;
+	Format: string;
+	DPI: string;
+	Generate: string;
 };
+
+export { english, french, finnish, german, swedish, spanish, Translation };
