+++ src/lib/local/sv.ts
@@ -1,0 +1,9 @@
+const translation = {
+	PageSize: 'Sidstorlek',
+	PageOrientation: 'Sidorientering',
+	Format: 'Format',
+	DPI: 'DPI',
+	Generate: 'Generera'
+};
+
+export default translation;
