+++ tsconfig.json
@@ -1,25 +1,25 @@
 {
-  "compilerOptions": {
-      "alwaysStrict": true,
-      "baseUrl": ".",
-      "declaration": true,
-      "experimentalDecorators": true,
-      "lib": [
-          "dom",
-          "es2020"
-      ],
-      "module": "commonjs",
-      "moduleResolution": "node",
-      "noImplicitThis": true,
-      "noUnusedLocals": true,
-      "outDir": "./dist/",
-      "preserveConstEnums": true,
-      "removeComments": true,
-      "strictFunctionTypes": true,
-      "strictNullChecks": true,
-      "sourceMap": true,
-      "target": "es6",
-      "esModuleInterop": true,
-  },
-  "include" : ["lib/*"] 
+	"compilerOptions": {
+		"alwaysStrict": true,
+		"baseUrl": ".",
+		"emitDeclarationOnly": true,
+		"declaration": true,
+		"declarationMap": true,
+		"declarationDir": "./dist/types",
+		"experimentalDecorators": true,
+		"lib": ["dom", "ESNext"],
+		"module": "ESNext",
+		"moduleResolution": "node",
+		"noImplicitThis": true,
+		"noUnusedLocals": true,
+		"outDir": "./dist/lib",
+		"preserveConstEnums": true,
+		"removeComments": true,
+		"strictFunctionTypes": true,
+		"strictNullChecks": true,
+		"sourceMap": true,
+		"target": "ESNext",
+		"esModuleInterop": true
+	},
+	"include": ["src/lib"]
 }
